# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Statera API (Partial Static Spec)
  version: 1.0.0
  description: >-
    Partial static OpenAPI contract for the Flask backend. It does not yet cover
    health probes, worker-health endpoints, feature flags, messages,
    notifications, categories, merchants, budgets, memorized transactions, or
    the open-banking route group.
  license:
    name: Proprietary
servers:
  - url: /
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
...
  responses:
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    OkEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/GenericObject'
        meta:
          $ref: '#/components/schemas/GenericObject'
      additionalProperties: true
      required:
        - ok
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
        errors:
          type: array
          items:
            type: string
        code:
          type: string
        error_code:
          type: string
        retry_after:
          type: integer
      additionalProperties: true
...
    RequestEmailChangeLinkRequest:
      type: object
      required:
        - new_email
        - current_password
      properties:
        new_email:
          type: string
          format: email
        current_password:
          type: string
    ConfirmEmailChangeRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    RequestPasswordChangeLinkRequest:
      type: object
      required:
        - current_password
      properties:
        current_password:
          type: string
    ConfirmPasswordChangeRequest:
      type: object
      required:
        - token
        - new_password
        - confirm_password
      properties:
        token:
          type: string
        new_password:
          type: string
          minLength: 8
        confirm_password:
          type: string
...
paths:
...
  /api/auth/profile/request-email-change-link:
    post:
      tags: [Auth]
      summary: Request an email change confirmation link.
      operationId: requestEmailChangeLink
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestEmailChangeLinkRequest'
      responses:
        '200':
          description: Verification link sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/auth/profile/confirm-email-change:
    post:
      tags: [Auth]
      summary: Confirm an email change token.
      operationId: confirmEmailChange
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmEmailChangeRequest'
      responses:
        '200':
          description: Email changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auth/profile/request-password-change-link:
    post:
      tags: [Auth]
      summary: Request a password change confirmation link.
      operationId: requestPasswordChangeLink
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordChangeLinkRequest'
      responses:
        '200':
          description: Password change link sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/auth/profile/confirm-password-change:
    post:
      tags: [Auth]
      summary: Confirm a password change token.
      operationId: confirmPasswordChange
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmPasswordChangeRequest'
      responses:
        '200':
          description: Password changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
...